requirements.txt
streamlit==1.35.0
transformers==4.41.0
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
PyMuPDF==1.24.3
requests==2.32.2
accelerate==0.30.0
sentencepiece==0.2.0
protobuf==4.25.3
einops==0.8.0
timm==1.0.3
pandas==2.2.2
rdflib==7.0.0
plotly==5.22.0
scipy==1.13.0
numpy==1.26.4
